pyproject.toml
[project]
name = "the-commerce-guild-assistant"
version = "0.1.0"
description = "A RAG-powered knowledge assistant for the My Time series (Sandrock & Portia) using multi-representation indexing. Optimizes retrieval of complex wiki structures and tables for high-context LLMs."
authors = [
    {name = "xilei-wang-90",email = "wangxilei1990@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "markdownify (>=1.2.2,<2.0.0)",
    "chromadb (>=1.5.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "langsmith (>=0.7.7,<0.8.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "sentence-transformers (>=5.3.0,<6.0.0)",
    "chainlit (>=2.10.0,<3.0.0) ; python_version >= \"3.12\" and python_full_version < \"4.0.0\"",
]


[tool.poetry]
packages = [
    {include = "guild_assistant", from = "src"},
    {include = "guild_assistant_web", from = "src"},
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.pytest.ini_options]
pythonpath = ["src", "."]
